pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ibmcloud-local"
version = "0.1.0"
description = "Local emulator for IBM Cloud services — like LocalStack, but for IBM Cloud"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "pydantic>=2.0",
    "httpx>=0.27.0",
    "click>=8.1.0",
    "rich>=13.0",
    "PyJWT[crypto]>=2.8",
    "cryptography>=41.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "ruff>=0.5.0",
    "httpx>=0.27.0",
]

[project.scripts]
ibmcloud-local = "cli.ibmcloud_local:cli"

[tool.setuptools.packages.find]
include = ["src*", "cli*"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"

[tool.ruff]
line-length = 100
